+++ docs/changelog.md
@@ -1,11 +1,16 @@
 # Changelog
 
+## (unreleased)
+
+- C4PlantUMLExporter: Fixes #83 (C4Plantuml export renders url property in wrong format).
+
 ## 1.17.0 (27th October 2023)
 
 - C4PlantUMLExporter: Adds support for border style and stroke width (#73).
 - C4PlantUMLExporter: Adds support for sequence diagrams (#66).
 - C4PlantUMLExporter: Relationships are now rendered as `Rel` rather than `Rel_D`.
 - MermaidDiagramExporter: Fixes #80 (Mermaid render error when description contains `(` character).
+- Updated dependencies and minimum Java version (17).
 
 ## 1.16.1 (11th August 2023)